+++ src/luamodalwindow.h
@@ -1,0 +1,43 @@
+// Copyright 2022 The Forgotten Server Authors. All rights reserved.
+// Use of this source code is governed by the GPL-2.0 License that can be found in the LICENSE file.
+
+#ifndef FS_LUAMODALWINDOW_H
+#define FS_LUAMODALWINDOW_H
+
+#include "luascript.h"
+
+class LuaModalWindow : public LuaScriptInterface
+{
+public:
+	static int luaModalWindowCreate(lua_State* L);
+	static int luaModalWindowDelete(lua_State* L);
+
+	static int luaModalWindowGetId(lua_State* L);
+	static int luaModalWindowGetTitle(lua_State* L);
+	static int luaModalWindowGetMessage(lua_State* L);
+
+	static int luaModalWindowSetTitle(lua_State* L);
+	static int luaModalWindowSetMessage(lua_State* L);
+
+	static int luaModalWindowGetButtonCount(lua_State* L);
+	static int luaModalWindowGetChoiceCount(lua_State* L);
+
+	static int luaModalWindowAddButton(lua_State* L);
+	static int luaModalWindowAddChoice(lua_State* L);
+
+	static int luaModalWindowGetDefaultEnterButton(lua_State* L);
+	static int luaModalWindowSetDefaultEnterButton(lua_State* L);
+
+	static int luaModalWindowGetDefaultEscapeButton(lua_State* L);
+	static int luaModalWindowSetDefaultEscapeButton(lua_State* L);
+
+	static int luaModalWindowHasPriority(lua_State* L);
+	static int luaModalWindowSetPriority(lua_State* L);
+
+	static int luaModalWindowSendToPlayer(lua_State* L);
+
+private:
+	friend class LuaScriptInterface;
+};
+
+#endif // FS_LUMODALWINDOW_H
